+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Launch finish time in case of fork-join launch, by @PavelSakharchuk
 ### Added
 - `rp.launch.uuid` property, to append to and existing launch without start/finish launch calls, by @HardNorth
+- Attribute values truncation and its properties, by @HardNorth
 
 ## [5.1.11]
 ### Added